lc22

+++ deleteTodo.js
@@ -1,7 +1,7 @@
 import { makeQuery } from './db.js'
 
-async function deleteTodo(id) {
-  const query = `DELETE FROM todos WHERE id=${id}`
+async function deleteTodo({ id, userId }) {
+  const query = `DELETE FROM todos WHERE id=${id} AND user_id=${userId}`
 
   await makeQuery(query)
